Separated posts and users requests by using routers.
